Working read from config files

+++ src/main/java/com/derongan/minecraft/looty/item/components/AreaComponent.java
@@ -3,13 +3,13 @@ package com.derongan.minecraft.looty.item.components;
 import com.badlogic.ashley.core.Component;
 
 public class AreaComponent implements Component {
-    public float radius;
+    public double radius;
 
-    private AreaComponent(float radius) {
+    private AreaComponent(double radius) {
         this.radius = radius;
     }
 
-    public static ComponentFactory create(float radius){
+    public static ComponentFactory create(double radius){
         return ()->new AreaComponent(radius);
     }
 }
